refactor: read article categories actions
