+++ src/actions/cms/articleCategories/read.ts
@@ -8,75 +8,15 @@ import type {
     ExpandedArticleCategoryWithCover,
     ArticleCategoryWithCover,
 } from '@/cms/articleCategories/Types'
+import { readArticleCategories, readArticleCategory } from '@/server/cms/articleCategories/read'
 
-export async function readArticleCategories(): Promise<ActionReturn<ArticleCategoryWithCover[]>> {
-    try {
-        const categories = await prisma.articleCategory.findMany({
-            include: {
-                articles: {
-                    take: 1,
-                    include: {
-                        coverImage: true
-                    }
-                },
-            },
-            orderBy: {
-                createdAt: 'desc'
-            }
-        })
-        const categoriesWithCover = await Promise.all(categories.map(async category => (
-            {
-                ...category,
-                coverImage: (await getCoverImage(category))
-            }
-        )))
-        return { success: true, data: categoriesWithCover }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function readArticleCategoriesAction(): Promise<ActionReturn<ArticleCategoryWithCover[]>> {
+    //TODO: only read categories that user has visibility
+    return await readArticleCategories()
 }
 
 
-export async function readArticleCategory(name: string): Promise<ActionReturn<ExpandedArticleCategoryWithCover>> {
-    try {
-        const category = await prisma.articleCategory.findUnique({
-            where: {
-                name
-            },
-            include: {
-                articles: {
-                    orderBy: {
-                        createdAt: 'desc'
-                    }
-                }
-            },
-        })
-        if (!category) return createActionError('NOT FOUND', `Category ${name} not found`)
-        const categoryWithCover = {
-            ...category,
-            coverImage: await getCoverImage(category)
-        }
-        return { success: true, data: categoryWithCover }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
-}
-
-/**
- * Get cover image for article category
- * Returns coverImage of a article in the category
- */
-async function getCoverImage(category: ExpandedArticleCategory): Promise<Image | null> {
-    if (category.articles.length === 0) return null
-    const coverImage = await prisma.cmsImage.findUnique({
-        where: {
-            id: category.articles[0].coverImageId
-        },
-        include: {
-            image: true
-        }
-    })
-    if (!coverImage) return null
-    if (!coverImage.image) return null
-    return coverImage.image
+export async function readArticleCategoryAction(name: string): Promise<ActionReturn<ExpandedArticleCategoryWithCover>> {
+    //TODO: only read if right visibility
+    return await readArticleCategory(name)
 }
