+++ src/app/articles/[category]/layout.tsx
@@ -1,6 +1,6 @@
 import styles from './layout.module.scss'
 import SideBar from './SideBar'
-import { readArticleCategory } from '@/cms/articleCategories/read'
+import { readArticleCategoryAction } from '@/cms/articleCategories/read'
 import { notFound } from 'next/navigation'
 import type { ReactNode } from 'react'
 
@@ -13,7 +13,7 @@ type PropTypes = {
 
 export default async function ArticleCategoryLayout({ params, children }: PropTypes) {
     const categoryName = decodeURIComponent(params.category)
-    const res = await readArticleCategory(categoryName)
+    const res = await readArticleCategoryAction(categoryName)
     if (!res.success) return notFound()
     const category = res.data
